Revert "Use better Histogram for metrics"
This reverts commit 0bf4a915f0ba371e2a07fb22f3af3a25eae7c06a.

+++ src/gas_pool/gas_pool_core.rs
@@ -72,7 +72,7 @@ impl GasPool {
             .await?;
         self.metrics
             .reserved_gas_coin_count_per_request
-            .observe(gas_coins.len() as f64);
+            .observe(gas_coins.len() as u64);
         Ok((
             sponsor,
             reservation_id,
@@ -190,7 +190,7 @@ impl GasPool {
         let elapsed = cur_time.elapsed().as_millis();
         self.metrics
             .transaction_execution_latency_ms
-            .observe(elapsed as f64);
+            .observe(elapsed as u64);
         let net_gas_usage = effects.gas_cost_summary().net_gas_usage();
         let new_daily_usage = self.gas_usage_cap.update_usage(net_gas_usage).await;
         self.metrics
